backend/requirements.txt
asgiref==3.9.1
cfgv==3.4.0
distlib==0.3.9
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.12
inflection==0.5.1
nodeenv==1.9.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.2.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
virtualenv==20.31.2
